feat: allow video upload timeout to be set, fix other kraken endpoints

+++ rest-kraken/src/main/java/com/github/twitch4j/kraken/domain/KrakenVideo.java
@@ -22,7 +22,7 @@ public class KrakenVideo {
     private String title;
     private String description;
     private String descriptionHtml;
-    private Integer broadcastId;
+    private Long broadcastId;
     private String broadcastType;
     private String status;
     private String tagList;
